Refactor file structure for snapshots folder
The idea is that you should be able to add the baseline snapshots to a
git/other version control system index. In order to do that, I want to
make the filenames of the baseline images really simple. I decided to go
with "baseline.png" for the baseline image, "candidate.png" for a new
snapshot that doesn't match the baseline, and "diff.png" for the diff
representation.

Removing the timestamp meant that fewer files will be created.