Move viewport_width config to individual examples
Instead of having a global configuration for the screen sizes we should
screenshot for components, we now allow individual components to define
their own set of sizes. This is more reasonable, since many components
won't be responsive at all, and then rendering in different screen sizes
is not necessary.

I've removed the config and added a way to specify a third argument when
describing a component: an array of screen widths. Here's an example:

  likadan.define('foo', renderFoo, [320, 1024])

The default size is 1024 (somewhat arbitrarily picked), so if you don't
specify sizes, 1024 will be used.