Remove the need for passing in an `out` element
Before, examples would be passed an `out` element that they were meant
to render into. This was weird, and many people (among them @lencioni)
asked if we couldn't just return the element instead. The reason I had
that there was because of how React renders into the DOM. I worked
around it a bit by letting examples return an element (or a rendered
React component) instead.

Please note that the dependency to React introduced is merely a "soft"
dependency. We only look for a `getDOMNode` method on what's returned. I
don't want this project to be React specific so I decided not to bring
in the full React lib as a dependency.