Add basic page with list of diffs
I want a quick way to review all diffs and approve/reject them. To do
that, I'm adding a `likadan review` tool that will spin up a server,
open a browser window and point it to /review. This url will serve a UI
that you can use to approve/reject/delete diffs. But for now it just
lists the diff images.