Clear out all visible elements in between test runs
This is a bit hacky, but here's the rationale:
In the Brigade codebase, we have a component (Dialog) that renders
elements directly in the body, outside of the default component DOM
node (I believe this technique is called "portal component" or
something). Before this commit, rendering a dialog would leave unwanted
elements on the page that leaked into the next likadan example. The code
in this commit will find all top-level elements that aren't invisible
and remove them from the DOM. The reason we only remove visible ones is
that we (Brigade) have an invisible div containing icon svg xml that we
need for icons to work. If we remove it, the likadan run will fail.

I'm pretty sure I'll have to revisit this decision at some point.
Alternatives:

 - Clear the entire DOM in between test runs (this probably implies
   having some sort of `beforeEach` concept for likadan).
 - Let examples themselves control their own cleanup.