Improve error handling
Before, if an example failed to render, we would stop all execution and
silently exit. This was because `driver.execute_script` from within Ruby
doesn't handle errors properly. An error thrown while executing a script
from Ruby would yield an `undefined` result. This would in turn be
interpreted as a valid signal that we are done.

To fix this, I've made renderCurrent always return an object. If an
error occurs, we log it to the js console, and return the error string
to the Ruby code. I've also made sure that you can trigger a single
example from passing in a `name` parameter. This way it's easy to
re-render a single failing example and debug.