Remove DOM manipulation when rendering example
Before, we would take the rendered element and attach it to the
`#likadan-output` div. This approach had a few problems:

 - React don't want its elements to be moved around in the DOM. This
   project is definitely going to be used with React. One issue was that
   we didn't have a way to unmount rendered components.
 - It hurts performance slightly to move stuff around in the DOM.

I've changed this to instead allow examples to be rendered anywhere in
the DOM themselves. We remember the rendered element and unmount it (in
case of React) or remove it from the DOM (plain JS).

I've been trying to stay clear of React dependencies. This commit
makes it a bit more tied to React, but not completely. I might look at
an alternative solution to unmounting for instance. We could allow
before and after hooks to be defined. That way it could be up to the
project itself to unmount.