Allow .likadan.yaml to contain ERB
The main reason being that we need to be able to make the configuration
file dynamic (soon, there'll be secret information inside it). I've seen
other projects allow ERB in YAML (Rails being one, for database.yml).

I must admit I've been on the fence about this. It does introduce a very
powerful tool to the configuration. One that can be misused to execute
malicious code.

The alternative I can think of would be to allow those configuration
options that need to be dynamic to be either in the config file or in an
ENV variable. I might make this change later depending on the outcome of
this whole thing.