Add `likadan upload_diffs` command line tool
This command will upload all diff images to an Amazon S3 bucket. The
output will contain URLs to all uploaded diffs. The idea here is that
you can persist diffs somewhere globally.

In order to make the S3 integration work, you need to add two new
configuration options to .likadan.yaml:

  s3_access_key_id: <your access key id>
  s3_secret_access_key: <your secret access key>

This commit is in service of making Likadan easier to use in a CI
environment. `likadan upload_diffs` is meant to be used in a script
doing something like this:

  # Check out the previous commit
  git checkout HEAD^

  # Run a likadan sweep to generate baselines
  likadan

  # Check out the latest commit
  git checkout HEAD

  # Run a likadan sweep to diff against previous baselines
  likadan

  # Upload diffs to Amazon S3
  likadan upload_diffs