Allow loading a different configuration file
I'm working on making Likadan better suited to be integrated in a CI
environment. In order to get better control over what configuration is
used, I needed to be able to specify a different .likadan.yaml file to
load.

I decided to implement a quick-fix for now by using an environment
variable. I had a look at other projects (e.g. scss-lint) and it looks
like the most common thing is to pass in a `--config` option when
running the command line tool. This would be the first option to be
added, and I didn't want to pay the cost of having to implement better
cli argument parsing right now, so an env variable will have to do.