Add `likadan.fdefine` to focus on one example
To better be able to re-run a single likadan example, @lencioni
suggested we add an `fdefine` method. This is similar to what Jasmine
and others have done, and allows you to run `likadan` with only one
example being tested.

I decided to implement a simple version that only uses the last
`fdefine` for now. In the future, we might want to change this so that
multiple `fdefine`s can exist and likadan will run all of them.

Fixes #11