Fix `likadan clean` for missing snapshots folder
Running `likadan clean` when the snapshots folder was missing would lead
to an error (something like "No such file or directory"). Since it's
likely that `likadan clean` will be added to scripts, I want it to be
more robust. Adding a guard to make sure that the folder exists will
help.