Switch to named viewports
This commit introduces named viewports, configured in .likadan.yaml. The
idea is that you set up a list of screen sizes that you want to
snapshot, and then refer to those in your examples, e.g.

```yaml
  viewports:
    desktop:
      width: 1024
      height: 768
    mobile:
      width: 320
      height: 444
```

```javascript
  likadan.define('foo', function() {
    return renderSomething();
  }, { viewports: ['desktop', 'mobile'] })
```

The old way of configuring widths in the `viewportWidths` option is no
longer supported.