Add simple rspec test
This project has been lacking spec coverage for quite some time now.
I've been sort of held back by not knowing if the direction was right
(no need to invest if things change), and also by the fact that the tool
itself is quite complex. There's a server running in a separate thread,
we use Selenium to talk to a Firefox browser and everything has to be
stiched together using a precise needle and a strong thread. That's no
excuse for not testing the damn thing though, so here goes.

I decided to write a full integration test right from the beginning.
Real files, real browsers, and real likadan processes are running. The
downside is that it gets harder to stub out certain functionality. But
on the other hand, specs will be written very close to how they work in
the wild.

The first test will verify the core flow:

 1. Running likadan on a new example generates a baseline
 2. Running likadan on the same example again without change does
 nothing.
 3. Running likadan on the same example when changes have been made
 results in a diff.

I'm not sure this type of test can be run in a Travis environment, but
I'll give it a shot.

Performance isn't great which will become a problem further down the
road. But for now it's fine. Better to have a test than no test.