Set <body> min-height to 1px on runner document
I was using likadan recently and came across an interesting error:

> Invalid image size, width: 320, height: 0

I have an example that renders a component that happens to take up 0
height. Likadan is just fine with this on the first pass, but on the
second pass when comparing a new version against a baseline, it
blows up with this error.

To prevent this from happening, I am setting a min-height of 1px on the
<body> element. If I understand how this works correctly, this will
cause the snapshot to be at least 1px tall, so it should never store a
0px tall image and run into this problem.

We may want to implement a more thorough fix for this issue at some
point, but this is simple and should at least help people out now.

Fixes #23