Avoid creating multiple S3 buckets
We started seeing the following error when likadan runs in CI:

> You have attempted to create more buckets than allowed
> (S3::Error::TooManyBuckets)

I think we can avoid this by only creating one bucket instead of a new
bucket every time likadan runs. I simply moved the random bucket name to
be a random file name, and created a non-random bucket name.

I decided to switch from SecureRandom.hex to SecureRandom.uuid because
uuid seems more appropriate for this type of thing.

I considered making the filename for the diffs just be the MD5 hash of
the file, which would avoid us from uploading duplicate data if the two
files happened to be the same. However, I didn't think the little bit of
added complexity of finding if the file already exists was worth the
small gain of no duplicates at this time.