Prevent one dimensional snapshots
In 39daf344 I added a min-height to the <body> element to prevent
snapshots from having 0 height when nothing is rendered. I had assumed
this would work, but I assumed incorrectly. This is because likadan
crops the snapshot to fit just the rendered element. To fix this for
real, I am just setting a minimum height and width of 1 on the cropping.

Fixes #23