Retry connecting to webdriver once
We've been seeing the error "unable to obtain stable firefox connection
in 60 seconds" when running Likadan in CI maybe 25% of the time
(unscientific). It seems that perhaps this unstable connection happens
when a newer version of Firefox is installed than what webdriver knows
how to handle [0].

As a work-around, I am just adding some code to retry it once if this
happens.

Addresses #27

[0]: https://github.com/SeleniumHQ/selenium/issues/835