Improve compatibility with things that require asynchronicity
At Brigade We have a likadan for a react component that uses
`requestAnimationFrame` when it renders to set some state. The snapshot
seems to sometimes be taken before this happens, and sometimes after it
happens. It would be nice if there was a way to ensure that the snapshot
is always taken after the `requestAnimationFrame`.

Jasmine solves this by having the function take a callback argument and
requiring the spec to call it when it is done. I think that a similar
technique will work for us here.

I implemented this by making the runner use WebDriver's
`execute_async_script` which injects a callback argument that needs to
be called when the script is done executing. This function can take a
value that gets passed back to Rubyland which we can use in the same way
as before to generate the snapshot. This allows us to write examples
that look like:

```js
likadan.define('foo', function(done) {
  setTimeout(function() {
    var elem = document.createElement('div');
    elem.innerHTML = 'Foo';
    document.body.appendChild(elem);
    done(elem);
  });
})
```

Fixes #28