Add CI example to README
The reason I created Likadan in the first place was to make visual
diffing work in a CI environment. Since a while ago, we've been running
it in a Jenkins environment at work, and I wanted to share how it can be
done so that others can pick it up.

Fixes #4