Rename project to Diffux-CI
We think that the project will benefit by being under the Diffux
umbrella. For those unfamiliar with Diffux, it is a visual diffing tool
more targeted at web sites not built with reusable components. Full
pages are snapshotted and diffed, and there's a powerful web UI that
comes with great tools to simplify a manual diffing process. Likadan is
built upon Diffux internals, so having it be part of the broader Diffux
org makes sense.

I made these substitutions:

  In Ruby code, "Likadan" became "DiffuxCI"

  In the Readme, "Likadan" became "Diffux-CI"

  "likadan", mostly became "diffux_ci", except for;

  In js code, the `window.likadan` object was named `window.diffux`. I
  did this because I found the "CI" extension to be redundant.

Partially fixes #35.