Prevent error when rendering a example from query string
A hidden functionality of Diffux CI is that you can pass in a `name`
parameter when loading URL '/' to render a single example. E.g.

 /?name=FooComponent

This was still working, but you would see errors about missing
`doneFunc` in the js console logs. I'm adding a stubbed method to
resolve that.