Add `diffux_ci debug` command line tool
This will fire up a browser window listing all your defined examples.
When you click on one, that example will be rendered in isolation.

I thought about adding the viewports to the mix, so that each example is
listed with its viewports (see example below). But that turned out to be
quite complicated because I had to share stuff between the backend and
frontend. I'm not entirely sure about the value of this tool either, so
I'll leave viewports for later.

  <FooComponent> with bars [large, small]
  <FooComponent> without bars [large, small]
  <BarComponent> [small]

I did duplicate a bit of haml stuff to get the page set up. I intend to
fix that in a follow-up.

Fixes #25