Fail on examples that have the same description
If you have two diffux examples with the same description, there is
currently a weird behavior where it will sometimes use the first one and
sometimes use the second one. Instead, we should notice this and trigger
an error.

```
diffux.define('My component', function() { ... });
diffux.define('My component', function() { ... });
```

It looks like we are using the term "name" in our code, so I decided to
stay locally consistent. A follow up commit will convert these to
"description" since that more accurately describes what it is.

If we want to get really serious about these kinds of problems, we could
write an ESLint plugin and package it under the diffux organization for
people to use.

Finishes #40