Clean up some code
There were a lot of long lines and a few other RuboCop warnings. I
decided that it would be nice to clean up some of this to make it easier
to read the code and work with this project.

This is not intended to be a functional change in any way.