Avoid writing screenshots to disk until necessary
I think that generally we want to avoid disk i/o unless we actually need
it. I noticed that for every screenshot, we were writing it to disk,
cropping it, writing it to disk again, and then sometimes deleting it
and other times renaming it. We can actually avoid a lot of these things
by just waiting to write it to disk until we actually care to save it.

While I was at it, I improved some of the variable names to more
accurately describe their values.