Revert "Require tilt/erubis explicitly before Sinatra"
This ended up blowing up Travis CI and it isn't very clear to me why at
this time. I don't think we need to worry about the warning this was
meant to avoid at this time and I'd rather have the build be green, so I
am reverting the commit.

This reverts commit a3533f0173676c71ce08741dbad841ce9246b780.