Try connecting to Firefox 3 times
We've been seeing

> Unable to obtain stable Firefox connection in 60 seconds

and previously added some retry logic in 715d7c4e to cut down the number
of times that we have to manually retrigger our CI jobs. However, we
still see this every once in a while. To make this even more pleasant, I
think it might be a good idea to retry a third time.

Addresses #27