Move version to own file
I want to add a --version command, which this move will make easier.
While I was at it I considered moving things around and re-namespacing
them to be more conventional but since that is a more in-depth change
than I want to make at this time I decided to leave it for another day.