Make DiffuxCIUtils.construct_url return a string again
When I converted this function to use the URI class, I changed what it
returned from a string to a URI object. I don't want to change this, so
I am calling .to_s on it at the end.

While I was at it, I decided to write some tests for this to ensure that
it won't be adversely affected by future refactoring. To make it a
little easier to to test things like this, I moved the reading of the
config file to its own function that can be mocked out in our tests.