Run screenshot and image comparison in separate threads
There is a clean and easy division of work here. The first thread runs
Firefox, rendering the examples, taking screenshots, cropping them, and
pushing that onto a queue. The second thread reads from the queue,
comparing the cropped screenshot to the baseline and generating diff
images. This allows us to take more advantage of systems that have
multiple CPUs, which will speed up this process a little.

In an unscientific test I ran on my laptop on the Brigade diffux test
suite using `time diffux_ci`, this brought a clean run where there were
no visual diffs down from 194s to 170s. This is a ~12% improvement and I
imagine there might be even more savings for runs that have diffs. Since
we run through this twice for every commit in CI, I would expect this to
bring our CI job runtimes down by maybe a minute.

We could probably explore taking this technique a step further by
increasing the number of threads to match the number of CPUs or some
configurable concurrency value and dividing up the work that way, but
this seems like a good enough start for now.

I am not very familiar with threads in Ruby, so it is likely that I have
made some mistakes here, but it seems to run on my laptop so I feel
pretty comfortable with this change.