Prefer cheaper element.offsetParent in visibility check
We are clearing out all visible elements from the document between
renders. We check visibility by getting the element's computed style.
This operation is expensive, so we want to avoid it whenever we can.
Thankfully, there is an inexpensive method that we can use to partly
determine the visibility of some elements--if an element has an
offsetParent, we know that it does not have display none or fixed
positioning.

This is a small optimization, but since it is so simple it seems like
one worth making.

I considered memoizing the visibility check for elements to make this
even faster, but I worry about the styles of elements shifting on us and
them not getting removed properly.