Improve accuracy of element coordinates
We want to use elem.getBoundingClientRect() instead of offsetTop and its
ilk because elem.getBoundingClientRect() is more accurate and it also
takes CSS transformations and other things of that nature into account
whereas offsetTop and company do not.

Note that this method returns floats, so we need to round those off
to integers before returning.

Unfortunately, this method is a bit slower, so we might see a little
worse performance after this change. However, the accuracy is well worth
this tradeoff.