Update comment about getDOMNode deprecation
In React 0.14, React.findDOMNode was moved to ReactDOM.findDOMNode.
Updating this comment to stay current.

I'm starting to think that in the long run we'll just have to make this
tool entirely React agnostic and put the burden of unmounting components
on the test suite. To do this, we will probably have to add some notion
of beforeEach and afterEach.