Guard against cropping too large
In 3c6e673ea I switched from offsetTop etc. to use
el.getBoundingClientRect(). Since this new method returns floats for
some values, I decided to round them up to make sure that we aren't
clipping stuff that we actually care about. However, this sometimes
gives us values that are actually outside of the screenshot that we get
from webdriver and when we try to crop these, the operation fails and we
get an error message.

We can guard against this by making sure that the crop width and height
are within the bounds of the screenshot before cropping.

I think that this will also resolve #37.