Use :fast_rgba when saving images
In 07dce0fa2 we started using the :fast_rgb option when saving images.
This seemed to work a lot of the time, but in practice it seems that
there are some situations where the screenshot includes some
transparency.

This caused issues because we compare a screenshot that has not been
saved with :fast_rgb against a screenshot that has been saved with
:fast_rgb. Since transparent pixels are converted to #000000 when saving
this way, this caused any transparent pixel to show up as a diff.
However, in the image we saved, these pixels looked identical in the
before and after images, causing much confusion.

Always saving as RGBA, should prevent this from happening.