Avoid unnecessary window resizing
Every time we resize the browser window, we pay a cost. Over the course
of a test suite, this adds up to a fairly significant cost. We can avoid
most of this cost by grouping examples to render by viewport.

We considered approaching this from the direction of avoiding
re-rendering examples when resizing the viewport, but that caused some
weirdness--namely we saw screenshots being taken at the wrong sizes and
examples bleeding into other examples. In the end, this seems like a
bigger win anyway since React rendering seems faster than resizing a
window.

In my unscientific experiment, this seems to take the Brigade diffux run
down from ~50s to ~35s.

We could probably further optimize this by changing the structure of
what getAllExamples() to look more like what we want here. However, I'll
leave that for another time.