Remove dependency on React
There was a soft dependency to React in that you could return a React
component from your examples instead of a DOM node. This dependency was
problematic in that we expected a globally defined `React` variable to
exist. I'm removing this dependency and moving the responsibility to the
examples instead.

To have a way to clean up event listeners and such, I'm adding a
`cleanOutElement` method that you can override. For React, this is where
you should unmount components.

Fixes #47