Simplify clearing out document between examples
We were previously just removing any visible elements from the body,
which requires a little overhead to determine if the element is visible.
I believe we were doing this to avoid removing our script tags which
were in the body. I think that if we move the script tags to the head,
we can do this more simply.

In my unscientific test, this didn't seem to have much of an impact on
performance, but it is simpler so I think it is still worth doing.