WIP use hashes of images for basic comparison
The idea is that if we generate a hash of the image, we can read just
that hash off the filesystem instead of the full image, resulting in a
performance improvement.

I thought about choosing CityHash since it claims to be very fast (at
least on 64-bit processors). However, it turned out to be a pain to
install on a Mac, so I went with XXHash instead, which claims to be
faster than CityHash.

Unfortunately, this seems to slow down the process a lot, making a
baseline run on the Brigade test suite go from ~30s to ~120s.

Addresses #65