Use base64 encoding for folder names
A problem with using the filesystem directly as the storage for
snapshots and diffs is that we've had to escape example descriptions in
order to use them as valid folder names. In #40, @lencioni and I have
discussed using base64 encoding to use as folder names. The upside is
that we can keep the real description throughout the UI (no more
weird underscores on the review screen). The downside is that the
structure is now obfuscated to a certain degree. What would before look
like this

snapshots
  |- _Foo_
      |- @320
         |- baseline.png
  |- _Bar_
      |- @320
         |- baseline.png
         |- diff.png
         |- candidate.png

Will now look like this

snapshots
  |- PEZvbz4NCg==
      |- @320
         |- baseline.png
  |- PEJhcj4NCg==
      |- @320
         |- baseline.png
         |- diff.png
         |- candidate.png

This probably doesn't matter, as the folder structure is an
implementation detail and not something that we expose. But if we find
out that it's hard to work with, we can provide tools to improve the UX
while still keeping the base64 encoding.