Store results of diffux run in result_summary.yaml
This is a pre-requisite for implementing uploading baselines for new
examples, as described in #63.

We currently don't store anything differently in the file system when an
examples is new, or when there's no diff. To be able to upload new
baselines after a run, we need somewhere to store that distinction. I
considered the following alternatives:

A) Storing a meta tag on the PNG image with information about if it is
new or not.
B) Storing an extra file inside the example's folder, with information
about if it's new or not.
C) Storing a summary log from a diffux run.

A and B both have N+1 performance implications, so I decided to go with
C, which could potentially be used for other things as well.