Fix new snapshots not uploading if diffs missing
The latest feature addition to the upload_diffs command was to upload
new snapshots along with the diffs. But a bug made it so that they were
only uploaded if there were diffs as well. If there was only a new
example added, nothing was uploaded. This commit fixes that.