Extract Logger class
I was thinking about #54 some more and I was considering outputting a
list of all of the images that had diffs in the last run. I wonder if
this will often end up being too long to be useful. So instead I
wondered if there was anything we can do to make the list that it
currently outputs more scannable. I think that coloring the results that
have diffs will help some. To manage this coloring, I decided to first
extract our current logging to a Logger class. My next commit will add
coloring.