Allow additional directories to be specified in the config
I added a way for diffux to be configured to look for resources in other
directories. This allows the usage of custom fonts and resolves the issue
https://github.com/diffux/diffux_ci/issues/82. I also added a test to
check that diffux can find files in other directories specified in its
config.

The console showed that diffux was trying to find a custom font file
with the url 'localhost:4567/CircularAir-bold.otf', so I added a
catch-all route for any extra uncaught urls to look for files in other
directories specified in the config. I implemented the
`public_directories` as an array, although it can be implemented as a
hash. Since the hash is more complicated to implement and doesn't seem
immediately necessary, I went forward with the simpler array implementation.