Move files over from diffux-core gem
We have deprecated the original diffux project and so no longer need
diffux-core to be a separate project. To simplify things, I am moving
the relevant files from that gem into here. This is a pretty straight up
move, with references updated to be DiffuxCI instead of Diffux. I think
there is a little code that can be removed from here, but I wanted to do
that in a separate commit to make the code trail easier to follow.