Remove the cluster finder
We only use this in Diffux, not Diffux CI, so now that we've moved the
code over from diffux-core, we can simplify a little.

We may want to bring this back or something like it when we improve the
UI for reviewing diffs (#61), but for now I want to keep things simple.