Attempt to sort out Travis CI configuration
After moving this repo and renaming everything to happo, the Travis
build seems to be failing.

For some reason, it is running `bundle exec rake` instead of `bash
headless_rspec.sh`, which doesn't work because we don't have rake
available in this project.

I am hoping that by simplifying some of this, I can get things working
properly again. This should also have the nice side-effect of making it
easier to run specs in development with the same version of rspec by
using `bundle exec rspec`.