Remove approve/reject feature
We brought this idea over from the original diffux project. However, as
we've used this tool, we have discovered that this feature is rarely
used and generally not that useful in this context. Most of the time
this will be run in CI where having approve/reject is just a little
confusing. To keep this project focused and reduce the amount of code we
need to maintain, I have decided to remove this feature.

We can likely simplify how we store files and tweak some of the
interface copy, but I'm going to leave that for another commit.

Fixes #71.