Rotate out current and previous images
Now that we don't have approve/reject buttons, we need to rotate out
images so that things make sense when running this project repeatedly
locally.

To keep things fast, we need to do as little work as possible in here.
With that in mind, we've settled on the following algorithm.

1. Delete previous.png if it exists.
2. Compare the current snapshot in memory against current.png if it
   exists.
3. If there is a diff, move current.png to previous.png and write out
   diff.png to disk and the current snapshot to current.png.
4. If there is no diff, return, leaving the old current.png in place.

This allows us to avoid having to always write the current snapshot to
disk even when nothing has changed, which should be the majority of the
time.