Get single S3 bucket instead of all buckets
Asking for all buckets may lead to access denied errors if the IAM role or user
associated with the provided access key pair does not have permissions for all
buckets. Furthermore, we only need to retrieve the bucket specified by the
environment variable `S3_BUCKET_NAME` to put diff images into, and we should
definitely have access to `S3_BUCKET_NAME` with the provided S3 credentials.

For some reason, the `service.bucket("bucket_name")` method was not specified
in the README for the `s3` library, and the only example in the documentation
for finding a single bucket was `service.buckets.find("bucket_name")`. I will
submit a pull request to update the documentation there as well. Also, since
`service.bucket(...)` doesn't check whether the bucket with the given name
exists, we should use the provided `exists?` method from the library to check
whether the bucket actually exists.