Clean up catchall route
Now that we are using early returns in this codepath, we can flatten the
logic a bit to improve readability of the code. While I was at it, I
added some comments explaining what is going on here and why it is
written the way it is, to help future explorers.