Use relative paths instead of absolute paths for images
With this change, the image elements in the generated `index.html` for
`happo upload_diffs` will use relative paths instead of absolute paths
for their source urls.

In my case, I'm using a proxy to allow developers without access to the
S3 bucket containing the image and HTML files Happo creates to also see
the visual diffing images. I would like to use the generated
`index.html` to present the visual diffing images, but the browser would
not be able to get to the absolute paths leading to the S3 bucket for
the images. Using relative paths allows the images to be retrieved using
the proxy as well.