Fix for Firefox 48+
Firefox 48 was released today, causing Happo to stop working. When
running Happo locally, Firefox would open but it would just sit there
doing nothing until it timed out.

We determined that this was broken due to the new version of Firefox
that came out today and filed an issue on the Selenium WebDriver
project:

  https://github.com/SeleniumHQ/selenium/issues/2559

> Can't fix.
>
> In Firefox 48, Mozilla requires extensions to be signed. They have
> refused to sign Selenium's Firefox driver extension (webdriver.xpi).
> Feel free to contact them and/or log an issue for them at
> https://bugzilla.mozilla.org
>
> The way forward now is to use GeckoDriver. Which is now the default
> implementation used in 3.0-beta
>
> This issue has been known as a long time coming, I highly recommend
> pinning your automation to use Firefox 45 ESR until a release of using
> GeckoDriver/marionette is stable enough for your tests.

We might want to try to detect the Firefox version and only apply this
option when the version is 48+, require folks to opt in to the option,
or allow folks to opt out of the option, but I'm not really sure if any
of that is important or not yet. The simplest thing is to require
Firefox 48+ and geckodriver to be installed.

More info about geckodriver can be found at:

- https://github.com/mozilla/geckodriver
- https://developer.mozilla.org/en-US/docs/Mozilla/QA/Marionette/WebDriver

It looks like the right way to tell Travis to use Firefox 48 is via the
`addons` option, which I've done here to keep our builds running.

  https://docs.travis-ci.com/user/firefox/