Try to install geckodriver on Travis
My last attempt failed on Travis and I'm pretty sure it is because
geckodriver is not available. This attempts to install it before running
the specs.

I copied this from:

- https://github.com/jnicklas/capybara/blob/232646ca/.travis.yml
- https://github.com/jnicklas/capybara/commit/f2d0a306

With selenium-webdriver v2, Travis was failing on:

  Selenium::WebDriver::Firefox.driver_path

with:

> undefined method `driver_path=' for
> Selenium::WebDriver::Firefox:Module

I believe that this method might not be available until version 3 of the
gem so I tried bumping the dependency to 3.0.0.beta2, but that didn't
work for other reasons.

I cracked open the code, and it looks like maybe we need to use `.path`
instead of `.driver_path` in this version?